Changes after moving AnyEscapableString from BatchEuphoria to RoddyToolLib.

GitPrivacy: ddj5jBde2HgIJuEKozvalBtVYKBdqv0CycS2HSUvbNDYrE2L6HK6SsJ5rLV+o/LYJcFRCQ8tH7U= ++aVzOmMVoIo08q4AnOlQZTfVFaOPldgkwaiM1NUCxOalNjfJRsXeI8p/AULrWutPYs7YeSx8cA=
