+++ RoddyCore/test/src/de/dkfz/roddy/execution/jobs/cluster/pbs/PBSJobManagerTest.groovy
@@ -20,7 +20,7 @@ import org.junit.Test
 
 import static org.junit.Assert.assertEquals
 
-import static de.dkfz.roddy.execution.EscapableString.*
+import static de.dkfz.roddy.tools.EscapableString.*
 
 /**
  */
